Change apps to examples and move setup.json
See also: #97
