+++ examples/dbmgr.py
