+++ examples/logger.py
